Moved AWT-dependent controllers to a new package
This is a fresh start on the effort already started in some older branches,
to remove AWT dependencies from Controller and a suite of subclasses.
This will enable use of those subclasses in the Oculus, thus allowing the
VR app to replicate more of the vZome UI.  It will also allow similar
reuse in vZome Online, without requiring the outdated and unsupported
j4ts-awt-swing JSweet candy.

On this branch, I will replicate changes from branches that are too old
to merge, and that probably used a weak strategy for the refactoring.
On this branch, I plan to use a more scrupulously clean approach: every
class in the com.vzome.desktop.controller package must be fully AWT-free.
Furthermore, I'm going to start with the leaf controllers, to allow an
incremental approach that can adhere to this cleanliness guideline.
