+++ desktop/src/main/java/org/vorthmann/j3d/J3dComponentFactory.java
@@ -3,7 +3,7 @@
 package org.vorthmann.j3d;
 
 import com.vzome.core.render.Scene;
-import com.vzome.desktop.controller.RenderingViewer;
+import com.vzome.desktop.awt.RenderingViewer;
 
 public interface J3dComponentFactory
 {
