+++ desktop/src/main/java/org/vorthmann/zome/app/impl/JsonClientShim.java
@@ -23,9 +23,9 @@ import com.vzome.core.render.RenderedModel;
 import com.vzome.core.render.Scene;
 import com.vzome.core.viewing.Camera;
 import com.vzome.core.viewing.Lights;
-import com.vzome.desktop.controller.CameraController;
+import com.vzome.desktop.awt.CameraController;
+import com.vzome.desktop.awt.RenderingViewer;
 import com.vzome.desktop.controller.JsonClientRendering;
-import com.vzome.desktop.controller.RenderingViewer;
 
 public abstract class JsonClientShim implements JsonClientRendering.EventDispatcher
 {
