+++ desktop/src/main/java/org/vorthmann/zome/app/impl/RedSpinAnimation.java
@@ -4,7 +4,7 @@ import javax.vecmath.AxisAngle4f;
 import javax.vecmath.Quat4f;
 import javax.vecmath.Vector3f;
 
-import com.vzome.desktop.controller.CameraController;
+import com.vzome.desktop.awt.CameraController;
 
 public class RedSpinAnimation implements AnimationController
 {
