+++ desktop/src/main/java/org/vorthmann/zome/app/impl/SymmetryController.java
@@ -24,7 +24,7 @@ import com.vzome.core.math.symmetry.Direction;
 import com.vzome.core.math.symmetry.OrbitSet;
 import com.vzome.core.math.symmetry.Symmetry;
 import com.vzome.core.render.RenderedModel;
-import com.vzome.desktop.controller.CameraController;
+import com.vzome.desktop.awt.CameraController;
 
 public class SymmetryController extends DefaultController
 {
